debug: add a testdev to capture the exit code
Acrn-unit-test uses a testdev to indicate the end of a test. This patch adds the
I/O port of the testdev so that the VM running the unit test can be paused and
later reset/destroyed.

The idle loop now also kicks the console so that the hypervisor shell works
after the VM pauses.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>