REVERT ME: guest: avoid overwriting multiboot boot info
There can be cases in which the allocated address for trampoline overlaps with
where multiboot boot info is placed, leading to the corruption of the boot info.

This patch reverses the order when visiting e820 entries to work this around. A
proper way shall be designed to mitigate this in the future.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>